Feat: Add login required

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -71,10 +72,17 @@ def product_view(request, product_id):
     return render(request, "products/product_view.html", context)
 
 
+@login_required
 def add_product(request):
     """
     Add a product to the store thorug the product management
     """
+    if not request.user.is_superuser:
+        messages.error(
+            request, "Only admin can add products"
+        )
+        return redirect(reverse("home"))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -96,10 +104,17 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def update_product(request, product_id):
     """
     Update a product in the store page
     """
+    if not request.user.is_superuser:
+        messages.error(
+            request, "Only admin can update products"
+        )
+        return redirect(reverse("home"))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -124,10 +139,17 @@ def update_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """
     Delete a product from the store page
     """
+    if not request.user.is_superuser:
+        messages.error(
+            request, "Only admin can delete products"
+        )
+        return redirect(reverse("home"))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
